requirements.txt
mysqlclient
djangorestframework
serializers
django-filter
django-multiselectfield
python-decouple
